gateway/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
prometheus_client>=0.20
pydantic==2.7.4
grpcio>=1.56,<2
grpcio-tools>=1.56,<2
protobuf>=4.24,<5
